[Parse] Split ParserPosition declaration to dedicated header file
